+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed an issue where exceptions were no printed in the **format** while the *--verbose* flag is set.
 * Fixed an issue where *--assume-yes* flag did not work in the **format** command when running on a playbook without a `fromversion` field.
 * Fixed an issue where the **format** command would fail in case `conf.json` file was not found instead of skipping the update.
+* Fixed an issue where integration with v2 were recognised by the `name` field instead of the `display` field in the **validate** command.
 
 # 1.2.10
 * Added validation for approved content pack use-cases and tags.