+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 * Fixed an issue where the ***generate-docs*** command reset the enumeration of line numbering after an MD table.
 * Updated the **upload** command to support mappers.
+* Fixed an issue with support for Docker Desktop on Mac version 2.5.0+.
+* Added support for vulture and mypy linting when running without docker.
 
 # 1.2.10
 * Added validation for approved content pack use-cases and tags.