+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Updated the **upload** command to support mappers.
 * Fixed an issue where integration with v2 were recognised by the `name` field instead of the `display` field in the **validate** command.
 * Added a playbook validation to check if a task script exists in the id set in the **validate** command.
+* Added new integration category `File Integrity Management` in the **validate** command.
 
 # 1.2.10
 * Added validation for approved content pack use-cases and tags.