+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+
+
+# 1.2.11
 * Fixed an issue where the ***generate-docs*** command reset the enumeration of line numbering after an MD table.
 * Updated the **upload** command to support mappers.
 * Fixed an issue where exceptions were no printed in the **format** while the *--verbose* flag is set.